third_party/WebKit/LayoutTests/animations/display-none-cancel-computedstyle-expected.txt

Canceling an animation should not leave dirty bits on display none elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(child).borderWidth is "10px"
PASS getComputedStyle(child).borderWidth is "5px"
PASS successfullyParsed is true

TEST COMPLETE
